requirements.txt
together>=1.0.0
openai>=1.0.0
groq>=0.5.0
python-dotenv>=1.0.0
regex>=2023.0.0
pandas>=2.0.0
numpy>=1.24.0
